challenge/requirements.txt
Django==2.0.4
djangorestframework==3.8.2
mysqlclient==1.3.12
numpy==1.14.3
pytz==2018.4
scipy==1.0.1
